コードレビュー対応
- InfinispanNamedTask インターフェース削除
- InfinispanManagedTask からインターフェース削除
- InfinispanSerializableTask の独自メソッドを削除し、フレームワークが用意しているメソッドに変更
